# Conflicts:
#	examples/common/common_tools.cpp
#	kernel_selector/core/cache/gpu_compiler.cpp
#	kernel_selector/core/cache/primitive_db.h
#	kernel_selector/core/cl_kernels/roi_pooling_v1_ref.cl
#	src/gpu/activation_gpu.cpp
#	src/gpu/cache/gpu_compiler.cpp
#	src/gpu/cache/gpu_linker.cpp
#	src/gpu/cache/kernels/deconvolution_gpu_bfyx_oiyx.naive.cl
#	src/gpu/cache/kernels/deconvolution_gpu_bfyx_yxio.naive.cl
#	src/gpu/cache/kernels/eltwise_gpu_bfyx.cl
#	src/gpu/cache/kernels/pooling_gpu_average_offset.cl
#	src/gpu/cache/kernels/pooling_gpu_bfyx_average_offset.cl
#	src/gpu/cache/kernels/pooling_gpu_bfyx_max_offset.cl
#	src/gpu/cache/kernels/pooling_gpu_max_offset.cl
#	src/gpu/cache/persistent_cache.cpp
#	src/gpu/convolution_gpu.cpp
#	src/gpu/crop_gpu.cpp
#	src/gpu/eltwise_gpu.cpp
#	src/gpu/kernels_cache.h
#	src/gpu/lrn_gpu.cpp
#	src/gpu/normalize_gpu.cpp
#	src/gpu/pooling_gpu.cpp
#	src/gpu/reorder_gpu.cpp
#	src/gpu/roi_pooling_gpu.cpp
#	src/gpu/softmax_gpu.cpp
#	src/program.cpp
#	tests/test_cases/convolution_gpu_test.cpp
#	tests/test_cases/pooling_gpu_test.cpp

+++ src/gpu/concatenation_gpu.cpp
@@ -44,7 +44,7 @@ struct concatenation_gpu : typed_primitive_impl<concatenation>
     {
         auto context = outer.get_program().get_engine()->get_context();
 
-        const int inputs_count = static_cast<int>(outer.inputs_count());
+        const size_t inputs_count = outer.inputs_count();
 
         if (!outer.can_be_optimized())
         {
